+++ djpcms/tests/test_response.py
@@ -1,0 +1,25 @@
+from djpcms.test import TestCase
+from djpcms.conf import settings
+from djpcms.tests.testmodel.models import Strategy
+
+class TestDjpResponse(TestCase):
+    '''Test functions in DjpResponse instances.'''
+    
+    def testOwnPageSimple(self):
+        context = self.get('/')
+        self.assertTrue(context['djp'].has_own_page())
+        self.makepage('search',Strategy)
+        context = self.get('/strategies/')
+        self.assertTrue(context['djp'].has_own_page())
+        
+    def testOwnPageApplication(self):
+        self.login()
+        link = '<a href="/edit-content/strategies/add/">edit</a>'
+        self.makepage('search',Strategy)
+        response = self.get('/strategies/add/', response = True)
+        self.assertFalse(response.context['djp'].has_own_page())
+        self.assertTrue(link not in response.content)
+        self.makepage('add',Strategy)
+        response = self.get('/strategies/add/', response = True)
+        self.assertTrue(response.context['djp'].has_own_page())
+        self.assertTrue(link in response.content)